added QTNETWORK lib link to cmake, refactored piqsl to use QTcpServer and QTcpSockets, removed threading, switched to QSharedPointers
